pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "django-ormql"
license = "Apache-2.0"
license-files = ["LICENSE.txt"]
description = "ORM Query Layer"
readme = "README.md"
authors = [{ name = "Raphael Michel", email = "mail@raphaelmichel.de" }]
keywords = [
    "django",
    "sql",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.10"
dependencies = [
    "sqlglot[c]==29.0.1",
]
dynamic = ["version"]

[project.urls]
"Homepage" = "https://github.com/pretix/django-ormql"

[project.optional-dependencies]
dev = [
    "pytest==9.*",
    "django<6.0",
    "factory_boy==3.*",
]

[tool.setuptools.dynamic]
version = { attr = "django_ormql.version" }

[tool.setuptools.packages.find]
include = ["django_ormql*"]
